// src/app/shared/model/gametemplate.ts
export interface Gametemplate {
    gameStartDate: any;
    blindStartDate: any;
    blindDuration: string;
    big: number;
    small: number;
    second: string;
    winner: string;
    completed: boolean;
    started: boolean;
    gameRef: string;
    userRef: string;
    startingStack: number;
    winnerName?: string;
    secondName?: string;
    }